panel/pyproject.toml
[project]
name = "oxide-panel"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0",
    "aiosqlite>=0.20",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "bcrypt>=4.2",
    "pillow>=11.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
include = ["panel*"]
